requirements.txt
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
torch>=2.1.0
transformers>=4.41.0
python-dotenv==1.0.1
pypdf==4.2.0
tqdm==4.66.4
streamlit==1.35.0
requests
